refactoring

+++ mandelbrot/src/mandelbrot/MandelbrotUtils.java
@@ -1,10 +1,10 @@
 package mandelbrot;
 
 public class MandelbrotUtils {
-	public static int getColor(int iterations, int maxIterations) {
+	public static int getColor(long iterations, long maxIterations) {
 		int color = 0;
 		if (iterations > 0) {//near to the set
-			int rgb = (iterations*255)/maxIterations;
+			int rgb = (int) ((iterations*255)/maxIterations);
 			color = (rgb << 16) | (rgb << 8) | rgb;
 		}
